+++ packages/copilot-studio-direct-to-engine-chat-adapter/CHANGELOG.md
@@ -14,7 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
    - Fixed race conditions between give up my turn and post activity, by [@compulim](https://github.com/compulim), in PR [#65](https://github.com/compulim/conversational-ai-chat-sdk/pull/65)
 - (Experimental) Added experimental /subscribe endpoint, by [@compulim](https://github.com/compulim), in PR [#62](https://github.com/compulim/conversational-ai-chat-sdk/pull/62) and [#64](https://github.com/compulim/conversational-ai-chat-sdk/pull/64)
 - Added `*/*;q=0.8` to accept header, by [@compulim](https://github.com/compulim), in PR [#70](https://github.com/compulim/conversational-ai-chat-sdk/pull/70)
-- (Experimental) Added `createFetchArguments` for custom fetch calls, by [@compulim](https://github.com/compulim), in PR [#XXX](https://github.com/compulim/conversational-ai-chat-sdk/pull/XXX)
+- (Experimental) Added `createFetchArguments` for custom fetch calls, by [@compulim](https://github.com/compulim), in PR [#71](https://github.com/compulim/conversational-ai-chat-sdk/pull/71)
+- (Experimental) Added `resumeConversationId` to resume conversation, by [@compulim](https://github.com/compulim), in PR [#75](https://github.com/compulim/conversational-ai-chat-sdk/pull/75)
 
 ## [0.0.0] - 2023-08-19